Use localhost if on Docker for Mac
Modern versions of Docker for Mac do not map the ports directly to the
IP address. After some trial and error, and looking at the Docker for
Mac documentation, the recommended way is to use "localhost".

This commit adds a simple helper that can pretty reliably detect when
you're running Docker for Mac (and Docker for Windows, I'm not sure) and
tells the transport to use "localhost" instead of the IP.

Signed-off-by: Tom Duffield <tom@chef.io>