Allow building data_image behind proxy
Addresses #147

**User Story**
As a build engineer,
I want to use kitchen-dokken behind a proxy without direct access to the
  internet,
So that it may be used in corporate CI/CD pipelines.

**Test Case**
*GIVEN* instance executing kitchen does not have direct internet access
*WHEN* performing integration testing with kitchen-dokken
*EXPECT* the ability to pull Docker images via proxy
*EXPECT* the ability to control proxy access while building the baked-in
  data_image

Maybe not the best approach, but it provides the minimum flexibility needed to
perform testing in an environment with strict access control.