Copy chef_version into provisioner so license acceptance works correctly
I originally thought about deprecating the `chef_version` attribute in
the driver but instead opted to have this driver continue working like
customers expect. It is a weird case because it extends the ChefZero
provisioner but does not ever install chef - the driver selects a docker
image of chef to run. The license acceptance logic depends on
`product_version` being set to the correct version so we need to copy it
from the driver into the provisioner.

Signed-off-by: tyler-ball <tball@chef.io>