update travis-ci configuration
Pull request #195 changed the platform fedora to centos, so .travis.yml
needs to match. Additionally, Travis now offers some linting advice, and
it pointed out that 'sudo' is superfluous and that explicit values for
os, dist, and language should be used. Because the tests do a clone for
the local repository, shallow cloning needs to be turned off. Finally, I
updated the rvm version to 2.7.1.