Fix `docker_registry` plus some cleanups
This PR extends the feature added in #218. Unfortunately there's a million
little places where docker decides to reach out to docker.io and if you
already have that image, or if you're not over you quota, it's hard to catch
it. This time I've done a lot of work to make sure every case is caught.

Further, the 'data image' doesn't get built unless you have a remote docker
host which I didn't in my testing, so that never got properly tested and was
broken when used with `docker_registry`, but it now works.

In addition I abstracted out a bit of the repetative code used to assemble
image names as well fixed some blatantly false debug messages, and added some
additional exceptions we weren't catching, plus filled in places we were
catching anything blindly.

The code is easier to read now, and I'm quite confident it's working as
advertised as I'm using it to run tests.

Signed-off-by: Phil Dibowitz <phil@ipom.com>